Add files via upload
✅ Implementação dos Fluxos

📁 Categoria
Categoria.java (Model)

CategoriaRepository.java

CategoriaService.java

CategoriaController.java


📁 Palestrante
Palestrante.java (Model)

PalestranteRepository.java

PalestranteService.java

PalestranteController.java


📁 Evento
Evento.java (Model)

EventoRepository.java

EventoService.java

EventoController.java



📁 Inscrição
Inscricao.java (Model)

InscricaoRepository.java

InscricaoService.java

InscricaoController.java


🧩 Regras de Negócio – InscricaoService
❌ Impedir inscrições em eventos que já ocorreram.

🔁 Evitar inscrições duplicadas (mesmo usuário no mesmo evento).

👤 Restringir inscrições apenas a usuários do tipo ALUNO. Usuários do tipo FACULDADE não podem se inscrever.



🔄 Funcionalidades de EventoPalestrante
🔹 Endpoint 1
GET /eventos/palestrante/{id}/futuros

Retorna todos os eventos futuros em que o palestrante com o ID informado está confirmado.

🔹 Endpoint 2
GET /eventos/futuros-por-nome-palestrante

Retorna eventos futuros com base em uma busca parcial pelo nome do palestrante.

Parâmetro obrigatório:

nome: parte do nome do palestrante.

Parâmetros opcionais:

page, size, sort: suporte a paginação e ordenação por data.

Resultado:
Lista de eventos futuros com pelo menos um palestrante cujo nome contenha o valor informado.



🔍 Observação Técnica
A relação @ManyToMany foi substituída pela entidade intermediária EventoPalestranteID, permitindo consultas mais flexíveis e diretas, como eventos futuros de um palestrante.

🗂️ Arquivos relacionados:
EventoPalestrante.java (Model)

EventoPalestranteRepository.java (com método customizado)

EventoService.java (consome a consulta do repositório)

EventoController.java (exposição dos novos endpoints)



⚙️ Refatoração Técnica — Uso de Exceção Personalizada
🧵 Contexto
Antes, o método buscarPorId() em CategoriaService lançava uma exceção genérica (throws Exception) quando a categoria não era encontrada, o que resultava em resposta HTTP 500 — mesmo quando o erro era esperado (ex.: 404 Not Found).

🆕 Solução
Criação da classe RecursoNaoEncontradoException, que estende RuntimeException e utiliza a anotação @ResponseStatus(HttpStatus.NOT_FOUND).

✅ Benefícios
Retorno automático de HTTP 404 para recursos não encontrados.

API mais semântica e alinhada com boas práticas REST.

Maior clareza e previsibilidade no tratamento de erros.

Evita o uso genérico de throws Exception.

Facilita o tratamento global e personalizado de exceções.



🗄️ Estrutura Atualizada do Banco de Dados
Remoções
Campos dataInicio, dataFim, vagasDisponiveis e nome foram removidos por não existirem nas tabelas originais.

Script Atualizado
sql
Copiar
Editar
DROP TABLE IF EXISTS inscricoes;
DROP TABLE IF EXISTS evento_palestrante;
DROP TABLE IF EXISTS eventos;

CREATE DATABASE actiosdb;
USE actiosdb;

CREATE TABLE faculdades (
    id_faculdade INT AUTO_INCREMENT PRIMARY KEY,
    nome VARCHAR(150) NOT NULL,
    localizacao VARCHAR(255),
    site VARCHAR(255)
);

CREATE TABLE usuarios (
    id_usuario INT AUTO_INCREMENT PRIMARY KEY,
    nome VARCHAR(100) NOT NULL,
    email VARCHAR(100) UNIQUE NOT NULL,
    senha VARCHAR(255) NOT NULL,
    curso VARCHAR(100),
    faculdade_id INT,
    tipo ENUM('ALUNO', 'FACULDADE') DEFAULT 'ALUNO',
    data_cadastro TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
    FOREIGN KEY (faculdade_id) REFERENCES faculdades(id_faculdade)
);

CREATE TABLE palestrantes (
    id_palestrante INT AUTO_INCREMENT PRIMARY KEY,
    nome VARCHAR(100) NOT NULL,
    email VARCHAR(100) UNIQUE NOT NULL,
    telefone VARCHAR(20),
    biografia TEXT
);

CREATE TABLE categorias (
    id_categoria INT AUTO_INCREMENT PRIMARY KEY,
    nome VARCHAR(100) UNIQUE NOT NULL
);

CREATE TABLE eventos (
    id_evento INT AUTO_INCREMENT PRIMARY KEY,
    titulo VARCHAR(255) NOT NULL,
    descricao TEXT,
    data DATE NOT NULL,
    horario VARCHAR(50),
    local VARCHAR(255),
    categoria_id INT,
    faculdade_id INT,
    FOREIGN KEY (categoria_id) REFERENCES categorias(id_categoria),
    FOREIGN KEY (faculdade_id) REFERENCES faculdades(id_faculdade)
);

CREATE TABLE evento_palestrante (
    evento_id INT,
    palestrante_id INT,
    PRIMARY KEY (evento_id, palestrante_id),
    FOREIGN KEY (evento_id) REFERENCES eventos(id_evento),
    FOREIGN KEY (palestrante_id) REFERENCES palestrantes(id_palestrante)
);

CREATE TABLE inscricoes (
    id_inscricao INT AUTO_INCREMENT PRIMARY KEY,
    usuario_id INT,
    evento_id INT,
    numero_inscricao VARCHAR(50) UNIQUE NOT NULL,
    data_inscricao TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
    FOREIGN KEY (usuario_id) REFERENCES usuarios(id_usuario),
    FOREIGN KEY (evento_id) REFERENCES eventos(id_evento)
);

+++ actios-backend/src/main/java/br/com/actios/actios_backend/repositorys/FaculdadeRepository.java
@@ -3,5 +3,5 @@ package br.com.actios.actios_backend.repositorys;
 import br.com.actios.actios_backend.model.Faculdade;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FaculdadeRepository extends JpaRepository<Faculdade, Long> {
+public interface FaculdadeRepository extends JpaRepository<Faculdade, Integer> {
 }
